Allow [[:rest]] arguments signature
That's what you get when using method_missing. With this,
"def helper(a, b, c)" and def "helper(*args)" form are supported.

The mixed form "def helper(a, c, *args)" is still not supported as
arguments_for_signature method would become more complicated and there
is no clear need for this.