Update Rails 6.1 gemfile
Also remove platforms from gemfile to just install ruby platform. Saves
time in CI. Also uses the latest version of Rails 6.1 and dependencies.

I mostly did this however because I was having some Nokogiri
installation issues in CI and removing the Gemfile.lock and reinstalling
with only `ruby` as the platform resolved it