CI: Use one job to check if required specs pass
Configuring GitHub to have many specs is hard to maintain.
In fact, the admin needs to duplicate what is defined in the ci.yml spec
matrix. It should be quite a bit easier this way, just having one CI job that
depends on the success of the entire spec matrix.